+++ packages/http-client/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @sales-planner/http-client
 
+## 0.8.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @sales-planner/shared@0.8.0
+
 ## 0.8.0
 
 ### Minor Changes